feat: Jira 1차 반응시간 계산 구현 완료

+++ src/main/java/com/pbl/tasktoolintegration/jira/entity/JiraIssueResponse.java
@@ -27,5 +27,5 @@ public class JiraIssueResponse extends BaseEntity {
     private JiraUser jiraUser;
 
     @Column(name = "responseSecond")
-    private Integer responseSecond;
+    private Long responseSecond;
 }
